# edge_cloud_management_api/specification/openapi.yaml
...
openapi: 3.0.3
info:
  title: Open Exposure Gateway API
  version: 1.0.1-wip
  description: |
    Open Exposure Gateway API allows API consumers to manage the
    Life Cycle of an Application, Discover Edge Cloud Zones and request Network Resources.
    # Overview
    The reference scenario foresees a distributed Telco Edge Cloud where any
    Application Delevoper, known as an Application Provider, can host and
    deploy their application according to their specifications and operational
    criteria (e.g. within an specific geographical zone for data protection
    purposes, ensure a minimum QoS for the application clients, etc).
    Through Telco Edge Cloud services Developers around the globe can be
    benefit from the traditional Cloud strengths but expertise and advantages
    of the Mobile Network Operators offering to their users an evolved
    experience for XR, V2X, Holographic and other new services.

    # Introduction
    The Edge Application Management API provides capabilities for lifecycle
    management of application, instances and edge cloud zone discovery.
    Lifecycle Management allows Application Provider to onboard
    their application to the Edge Cloud Platform which do bookkeeping,
    resource validation and other pre-deployment operations.
    Application details can contain components network specification,
    package type (QCOW2, OVA, CONTAINER, HELM), operating system details and
...
    Once the application is available on the Edge Cloud
...

...

...

...

...

...

...
servers:
...
paths:
...
  /partner:
    post: 
      tags:
      - FederationManagement
      summary: Creates one direction federation with partner operator platform.
      # security:
      #   - oAuth2ClientCredentials:
      #     - fed-mgmt
      # security:
      #   - jwt: [ ]

      operationId: edge_cloud_management_api.controllers.federation_manager_controller.create_federation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FederationRequestData'
        required: true
      responses:
        "200":
          description: Federation meta-info request accepted
          headers:
            Location:
              description: "Contains the URI of the newly created resource, according\
                \ to the structure: {apiRoot}/operatorplatform/federation/v1/partner/{federationContextId}"
              required: true
              style: simple
              explode: false
              schema:
                type: string
            Accept-Encoding:
              description: "Accept-Encoding, described in IETF RFC 7694"
              style: simple
              explode: false
              schema:
                type: string
            Content-Encoding:
              description: "Content-Encoding, described in IETF RFC 7231"
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FederationResponseData'
        # "400":
        #   description: Bad request
        #   content:
        #     application/problem+json:
        #       schema:
        #         $ref: '#/components/schemas/ProblemDetails'
            
        # "404": 
        #   description: Unauthorized
        #   content:
        #     application/problem+json:
        #       schema:
        #         $ref: '#/components/schemas/ProblemDetails'
        # "408": 
        #   description: Timeout
        #   content:
        #     application/problem+json:
        #       schema:
        #         $ref: '#/components/schemas/ProblemDetails'        
        # "409":
        #   description: Conflict
        #   content:
        #     application/problem+json:
        #       schema:
        #         $ref: '#/components/schemas/ProblemDetails'
        # "500":
        #   description: Internal Server Error
        #   content:
        #     application/problem+json:
        #       schema:
        #         $ref: '#/components/schemas/ProblemDetails'
      # callbacks:
      #   onPartnerStatusEvent:
      #     '{$request.body#/partnerStatusLink }':
      #       post:
      #         requestBody:
      #           description: |
      #             OP uses this callback api to notify partner OP about change in federation status, federation metadata or offered zone details. Allowed combinations of objectType and operationType are
      #             - FEDERATION - STATUS: Status specified by parameter 'federationStatus'.
      #             - ZONES - STATUS: Status specified by parameter 'zoneStatus'.
      #             - ZONES - ADD: Use parameter 'addZones' to define add new zones
      #             - ZONES - REMOVE: Use parameter 'removeZones' to define remove zones.
      #             - EDGE_DISCOVERY_SERVICE - UPDATE: Use parameter 'edgeDiscoverySvcEndPoint' to specify new endpoints
      #             - LCM_SERVICE - UPDATE: Use parameter 'lcmSvcEndPoint' to specify new endpoints
      #             - MOBILE_NETWORK_CODES - ADD: Use parameter 'addMobileNetworkIds' to define new mobile network codes.
      #             - MOBILE_NETWORK_CODES - REMOVE: Use parameter 'removeMobileNetworkIds' to remove mobile network codes.
      #             - FIXED_NETWORK_CODES - ADD: Use parameter 'addFixedNetworkIds' to define new fixed network codes.
      #             - FIXED_NETWORK_CODES - REMOVE: Use parameter 'removeFixedNetworkIds' to remove fixed network codes.
      #             - SERVICE_APIS - ADD/REMOVE: Parameter Usage 'addServiceAPIs / removeServiceAPIs' to add or remove Service APIs support.
      #           content:
      #             application/json:
      #               schema:
      #                 required:
      #                 - federationContextId
      #                 - modificationDate
      #                 - objectType
      #                 - operationType
      #                 type: object
      #                 properties:
      #                   federationContextId:
      #                     $ref: '#/components/schemas/FederationIdentifier' 
      #                   objectType:
      #                     type: string
      #                     enum:
      #                     - FEDERATION
      #                     - ZONES
      #                     - EDGE_DISCOVERY_SERVICE
      #                     - LCM_SERVICE
      #                     - MOBILE_NETWORK_CODES
      #                     - FIXED_NETWORK_CODES
      #                     - SERVICE_APIS
      #                   operationType:
      #                     type: string
      #                     enum:
      #                     - STATUS
      #                     - UPDATE
      #                     - ADD
      #                     - REMOVE
      #                   edgeDiscoverySvcEndPoint:
      #                     $ref: '#/components/schemas/ServiceEndpoint'
      #                   lcmSvcEndPoint:
      #                     $ref: '#/components/schemas/ServiceEndpoint'
      #                   addMobileNetworkIds:
      #                     $ref: '#/components/schemas/MobileNetworkIds'
      #                   removeMobileNetworkIds:
      #                     $ref: '#/components/schemas/MobileNetworkIds'
      #                   addFixedNetworkIds:
      #                     $ref: '#/components/schemas/FixedNetworkIds'
      #                   removeFixedNetworkIds:
      #                     $ref: '#/components/schemas/FixedNetworkIds'
      #                   addZones:
      #                     minItems: 1
      #                     type: array
      #                     description: "List of zones, which the operator platform\
      #                       \ wishes to make available to developers/ISVs of requesting\
      #                       \ operator platform."
      #                     items:
      #                       $ref: '#/components/schemas/ZoneDetails'
      #                   removeZones:
      #                     minItems: 1
      #                     type: array
      #                     description: "List of zones, which the operator platform\
      #                       \ no longer wishes to share."
      #                     items:
      #                       $ref: '#/components/schemas/ZoneIdentifier'
      #                   addServiceAPIs:
      #                     $ref: '#/components/schemas/serviceAPINames'
      #                   removeServiceAPIs:
      #                     $ref: '#/components/schemas/serviceAPINames'  
      #                   zoneStatus:
      #                     minItems: 1
      #                     type: array
      #                     items:
      #                       required:
      #                       - status
      #                       - zoneId
      #                       type: object
      #                       properties:
      #                         zoneId:
      #                           $ref: '#/components/schemas/ZoneIdentifier'
      #                         status:
      #                           $ref: '#/components/schemas/Status'
      #                   federationStatus:
      #                     $ref: '#/components/schemas/Status'
      #                   modificationDate:
      #                     type: string
      #                     description: Date and time of the federation modification
      #                       by the originating partner OP
      #                     format: date-time
              
      #         responses:
      #           "204":
      #             description: Expected response to a successful call back processing
      #           "400":
      #             description: Bad request
      #             content:
      #               application/problem+json:
      #                 schema:
      #                   $ref: '#/components/schemas/ProblemDetails'
      #           "401":
      #             description: Unauthorized
      #             content:
      #               application/problem+json:
      #                 schema:
      #                   $ref: '#/components/schemas/ProblemDetails'
      #           "404":
      #             description: Not Found
      #             content:
      #               application/problem+json:
      #                 schema:
      #                   $ref: '#/components/schemas/ProblemDetails'
      #           "409":
      #             description: Conflict
      #             content:
      #               application/problem+json:
      #                 schema:
      #                   $ref: '#/components/schemas/ProblemDetails'
      #           "422":
      #             description: Unprocessable Entity
      #             content:
      #               application/problem+json:
      #                 schema:
      #                   $ref: '#/components/schemas/ProblemDetails'
      #           "500":
      #             description: Internal Server Error
      #             content:
      #               application/problem+json:
      #                 schema:
      #                   $ref: '#/components/schemas/ProblemDetails' 
      #           "503":
      #             description: Service Unavailable
      #             content:
      #               application/problem+json:
      #                 schema:
      #                   $ref: '#/components/schemas/ProblemDetails'
      #           "520":
      #             description: Web Server Returned an Unknown Error
      #             content:
      #               application/problem+json:
      #                 schema:
      #                   $ref: '#/components/schemas/ProblemDetails'
      #           default:
      #             description: Generic Error
  /{federationContextId}/partner:
   get:
      tags:
      - FederationManagement
      summary: "Retrieves details about the federation context with the partner OP.\
        \ The response shall provide info about the zones offered by the partner,\
        \ partner OP network codes, information about edge discovery and LCM service\
        \ etc."
      # security:
      #   - jwt: [ ]
      operationId: edge_cloud_management_api.controllers.federation_manager_controller.get_federation
      parameters:
      - name: federationContextId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/FederationContextId'
      responses:
        "200":
          description: Federation meta-info request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        # "401":
        #   description: Unauthorized
        #   content:
        #     application/problem+json:
        #       schema:
        #         $ref: '#/components/schemas/ProblemDetails'
        # "404":
        #   description: Not Found
        #   content:
        #     application/problem+json:
        #       schema:
        #         $ref: '#/components/schemas/ProblemDetails'
        # "500":
        #   description: Internal Server Error
        #   content:
        #     application/problem+json:
        #       schema:
        #         $ref: '#/components/schemas/ProblemDetails'
   delete:
     tags:
     - FederationManagement
     summary: Remove existing federation with the partner OP
    #  security:
    #     - jwt: [ ]
     operationId: edge_cloud_management_api.controllers.federation_manager_controller.delete_federation
     parameters:
     - name: federationContextId
       in: path
       required: true
       style: simple
       explode: false
       schema:
         $ref: '#/components/schemas/FederationContextId'
     responses:
       "200":
         description: Federation removed successfully
      #  "401":
      #     description: Unauthorized
      #     content:
      #       application/problem+json:
      #         schema:
      #           $ref: '#/components/schemas/ProblemDetails'
      #  "404":
      #     description: Not Found
      #     content:
      #       application/problem+json:
      #         schema:
      #           $ref: '#/components/schemas/ProblemDetails'         
      #  "500":
      #     description: Internal Server Error
      #     content:
      #       application/problem+json:
      #         schema:
      #           $ref: '#/components/schemas/ProblemDetails'
...
components:
...
  schemas:
...
    CountryCode:
      pattern: "^[A-Z]{2}$"
      type: string
      description: ISO 3166-1 Alpha-2 code for the country of Partner operator
...
    CallbackCredentials:
      required:
      - clientId
      - clientSecret
      - tokenUrl
      type: object
      properties:
        tokenUrl:
          $ref: '#/components/schemas/Uri'
        clientId:
          type: string
          description: Client id for oauth2 client credentials flow.
        clientSecret:
          type: string
          description: Client secret for oauth2 client credentials flow.
      description: "Authentication credentials for callbacks. Callbacks use the same\
        \ security scheme, flows, and scopes as the forward path."
    GeoLocation:
      pattern: "^([-+]?)([\\d]{1,2})((((\\.)([\\d]{1,4}))?(,)))(([-+]?)([\\d]{1,3})((\\\
        .)([\\d]{1,4}))?)$"
      type: string
      description: "Latitude,Longitude as decimal fraction up to 4 digit precision"
...
    inline_response_200_1:
      required:
      - edgeDiscoveryServiceEndPoint
      - lcmServiceEndPoint
      type: object
      properties:
        edgeDiscoveryServiceEndPoint:
          $ref: '#/components/schemas/ServiceEndpoint'
        lcmServiceEndPoint:
          $ref: '#/components/schemas/ServiceEndpoint'
        allowedMobileNetworkIds:
          $ref: '#/components/schemas/MobileNetworkIds'
        allowedFixedNetworkIds:
          $ref: '#/components/schemas/FixedNetworkIds'
        offeredAvailabilityZones:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/ZoneDetails'
      example:
        allowedFixedNetworkIds:
        - allowedFixedNetworkIds
        - allowedFixedNetworkIds
        offeredAvailabilityZones:
        - geographyDetails: geographyDetails
          zoneId: zoneId
          geolocation: geolocation
        - geographyDetails: geographyDetails
          zoneId: zoneId
          geolocation: geolocation
        lcmServiceEndPoint: null
        allowedMobileNetworkIds:
          mncs:
          - mncs
          - mncs
          mcc: mcc
        edgeDiscoveryServiceEndPoint:
          ipv4Addresses:
          - 198.51.100.1
          - 198.51.100.1
          port: 0
          fqdn: fqdn
          ipv6Addresses:
          - 2001:db8:85a3::8a2e:370:7334
          - 2001:db8:85a3::8a2e:370:7334
...
    FederationIdentifier:
      pattern: "^[A-Za-z0-9][A-Za-z0-9-]*$"
      type: string
      description: Globally unique identifier allocated to an operator platform. This
        is valid and used only in context of MEC federation interface.
    FederationRequestData:
      required:
      - initialDate
      - origOPFederationId
      - partnerStatusLink
      type: object
      properties:
        origOPFederationId:
          $ref: '#/components/schemas/FederationIdentifier'
        origOPCountryCode:
          $ref: '#/components/schemas/CountryCode'
        origOPMobileNetworkCodes:
          $ref: '#/components/schemas/MobileNetworkIds'
        origOPFixedNetworkCodes:
          $ref: '#/components/schemas/FixedNetworkIds'
        initialDate:
          type: string
          description: Time zone info of the federation initiated by the originating
            OP
          format: date-time
        partnerStatusLink:
          $ref: '#/components/schemas/Uri'
        partnerCallbackCredentials:
          $ref: '#/components/schemas/CallbackCredentials'
    FederationResponseData:
      required:
      - federationContextId
      - partnerOPFederationId
      - platformCaps
      type: object
      properties:
        partnerOPFederationId:
          $ref: '#/components/schemas/FederationIdentifier'
        partnerOPCountryCode:
          $ref: '#/components/schemas/CountryCode'
        federationContextId:
          $ref: '#/components/schemas/FederationContextId'
        edgeDiscoveryServiceEndPoint:
          $ref: '#/components/schemas/ServiceEndpoint'
        lcmServiceEndPoint:
          $ref: '#/components/schemas/ServiceEndpoint'
        partnerOPMobileNetworkCodes:
          $ref: '#/components/schemas/MobileNetworkIds'
        partnerOPFixedNetworkCodes:
          $ref: '#/components/schemas/FixedNetworkIds'
        offeredAvailabilityZones:
          minItems: 1
          type: array
          description: "List of zones, which the operator platform wishes to make\
            \ available to developers/ISVs of requesting operator platform."
          items:
            $ref: '#/components/schemas/ZoneDetails'
    FederationContextId:
      pattern: "^[A-Za-z0-9][A-Za-z0-9-]*$"
      type: string
      description: This identifier shall be provided by the partner OP on successful
        verification and validation of the federation create request and is used by
        partner op to identify this newly created federation context. Originating
        OP shall provide this identifier in any subsequent request towards the partner
        op.
      readOnly: true
    ProblemDetails:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        cause:
          type: string
        invalidParams:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/InvalidParam'
    ServiceEndpoint:
      required:
      - port
      type: object
      properties:
        port:
          $ref: '#/components/schemas/Port'
        fqdn:
          $ref: '#/components/schemas/Fqdn'
        ipv4Addresses:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Ipv4Addr'
        ipv6Addresses:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Ipv6Addr'
      example:
        ipv4Addresses:
        - 198.51.100.1
        - 198.51.100.1
        port: 0
        fqdn: fqdn
        ipv6Addresses:
        - 2001:db8:85a3::8a2e:370:7334
        - 2001:db8:85a3::8a2e:370:7334
      anyOf:
      - required:
        - fqdn
      - required:
        - ipv4Addresses
      - required:
        - ipv6Addresses
    InvalidParam:
      required:
      - param
      type: object
      properties:
        param:
          type: string
        reason:
          type: string
    ZoneDetails:
      required:
      - geographyDetails
      - geolocation
      - zoneId
      type: object
      properties:
        zoneId:
          $ref: '#/components/schemas/ZoneIdentifier'
        geolocation:
          $ref: '#/components/schemas/GeoLocation'
        geographyDetails:
          type: string
          description: "Details about cities or state covered by the edge. Details\
            \ about the type of locality for eg rural, urban, industrial etc. This\
            \ information is defined in human readable form."
      example:
        geographyDetails: geographyDetails
        zoneId: zoneId
        geolocation: geolocation
    ZoneIdentifier:
      pattern: "^[A-Za-z0-9][A-Za-z0-9-]*$"
      type: string
      description: Human readable name of the zone.
...
    FixedNetworkIds:
      minItems: 1
      type: array
      description: List of network identifier associated with the fixed line network
        of the operator platform.
      items:
        type: string    
    Fqdn:
      type: string
      description: |
        Full qualified domain name of an application instance
...
    Mcc:
      pattern: "^\\d{3}$"
      type: string
    Mnc:
      pattern: "^\\d{2,3}$"
    MobileNetworkIds:
      type: object
      properties:
        mcc:
          $ref: '#/components/schemas/Mcc'
        mncs:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Mnc'
      example:
        mncs:
        - mncs
        - mncs
        mcc: mcc
...
    Ipv4Addr:
      type: string
      format: ipv4
      example: "192.168.0.1"
      description: |
        IP of the device. A single IPv4 address may be specified in
        dotted-quad form 1.2.3.4. Only this exact IP number will match the flow
        control rule.
    Ipv6Addr:
      type: string
      format: ipv6
      example: "2001:db8:85a3:8d3:1319:8a2e:370:7344"
      description: |
        IP of the device. A single IPv6 address, following IETF 5952
        format, may be specified like 2001:db8:85a3:8d3:1319:8a2e:370:7344
...
    Port:
      type: integer
      description: Port to stablish the connection
      minimum: 0
...
    serviceAPINames:
      minItems: 1
      type: array
      description: "List of Service API capability names an OP supports and offers\
        \ to other OPs \"quality_on_demand\", \"device_location\" etc."
      items:
        type: string
        enum:
        - QualityOnDemand
        - DeviceLocation
        - DeviceStatus
        - SimSwap
        - NumberVerification
        - DeviceIdentifier
    Status:
      type: string
      enum:
      - FAILED
      - TEMPORARY_FAILURE
      - AVAILABLE
      - LOCKED
      - NOT_AVAILABLE
    Uri:
      type: string
      example: https://charts.bitnami.com/bitnami/helm/example-chart:0.1.0
      description: |
        A Uniform Resource Identifier (URI) as per RFC 3986,
        identifies the endpoint within an Edge Cloud Zone where the user
        equipment may connect to the selected application instance
...